moved rename_input_files.py and config template to designated subdirectory in tools finalized rename_input_files.py modified config_template parameters and added comments
